完成所有编码

+++ bortus-manager/src/main/java/com/orvillex/bortus/manager/modules/scheduler/repository/JobInfoRepository.java
@@ -16,7 +16,7 @@ import org.springframework.data.jpa.repository.Query;
  */
 public interface JobInfoRepository extends JpaRepository<JobInfo, Long>, JpaSpecificationExecutor<JobInfo> {
     
-    List<JobInfo> findByJobGroup(Integer jobGroup);
+    List<JobInfo> findByJobGroup(Long jobGroup);
 
     @Query(value = "SELECT m.* FROM sys_job_info AS m WHERE m.trigger_status = 1 AND t.trigger_next_time <= ?1 ORDER BY info_id ASC LIMIT ?2", nativeQuery = true)
     List<JobInfo> scheduleJobQuery(Long maxNextTime, Integer pageSize);
