+++ bortus-manager/src/main/java/com/orvillex/bortus/manager/modules/scheduler/service/impl/JobInfoServiceImpl.java
@@ -57,7 +57,7 @@ public class JobInfoServiceImpl implements JobInfoService {
     }
 
     @Override
-    public List<JobInfo> findByJobGroup(Integer jobGroup) {
+    public List<JobInfo> findByJobGroup(Long jobGroup) {
         return jobInfoRepository.findByJobGroup(jobGroup);
     }
